1.9.29 Update

+++ src/main/java/net/rayfall/eyesniper2/skrayfall/v1_11/EffTitleV1_11.java
@@ -63,7 +63,7 @@ public class EffTitleV1_11 extends Effect {
     protected void execute(Event evt) {
         int timeTick = 60;
         if (time != null) {
-            timeTick = time.getSingle(evt).getTicks();
+            timeTick = (int) time.getSingle(evt).getTicks();
         }
 
         for (Player p : player.getAll(evt)) {
@@ -71,15 +71,15 @@ public class EffTitleV1_11 extends Effect {
 
             if (fadeIn != null && fadeOut != null) {
                 PacketPlayOutTitle packetPlayOutTimes = new PacketPlayOutTitle(EnumTitleAction.TIMES, null,
-                        fadeIn.getSingle(evt).getTicks(), timeTick, fadeOut.getSingle(evt).getTicks());
+                        (int) fadeIn.getSingle(evt).getTicks(), timeTick, (int) fadeOut.getSingle(evt).getTicks());
                 connection.sendPacket(packetPlayOutTimes);
             } else if (fadeIn == null && fadeOut != null) {
                 PacketPlayOutTitle packetPlayOutTimes = new PacketPlayOutTitle(EnumTitleAction.TIMES, null,
-                        5, timeTick, fadeOut.getSingle(evt).getTicks());
+                        5, timeTick, (int) fadeOut.getSingle(evt).getTicks());
                 connection.sendPacket(packetPlayOutTimes);
             } else if (fadeIn != null && fadeOut == null) {
                 PacketPlayOutTitle packetPlayOutTimes = new PacketPlayOutTitle(EnumTitleAction.TIMES, null,
-                        fadeIn.getSingle(evt).getTicks(), timeTick, 5);
+                        (int) fadeIn.getSingle(evt).getTicks(), timeTick, 5);
                 connection.sendPacket(packetPlayOutTimes);
             } else {
                 PacketPlayOutTitle packetPlayOutTimes =
